requirements.txt
dj-database-url==0.3.0
Django==1.6.11
django-appconf==1.0.1
django-classy-tags==0.6.2
django-cms==3.1.3
django-reversion==1.8.5
django-sekizai==0.8.2
Django-Select2==4.3.2
django-treebeard==3.0
djangocms-admin-style==1.0.4
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.3
djangocms-inherit==0.1
djangocms-installer==0.8.1
djangocms-link==1.7.1
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.7.0
djangocms-video==0.1
html5lib==0.9999999
Pillow==3.0.0
pytz==2015.7
six==1.10.0
South>0.8.4
tzlocal==1.2
wheel==0.24.0
